Refactored helpers to use index to export one object of helpers
