+++ db/index.js
@@ -11,6 +11,6 @@ mongoose.Promise = Promise;
 const db = mongoose.connection;
 
 db.on( 'error', console.error.bind( console, 'connection error:') );
-db.once( 'open', () =>{ console.log( `connected to mongoDB : "${ database }"!` ); });
+db.once( 'open', () =>{ console.log( `connected to "${ process.env.MONGO_URL }"!` ); });
 
 module.exports = db;
