+++ server/helpers/reservations.js
@@ -1,0 +1,24 @@
+const reformat = require('./reformat');
+
+
+const formatReservation = (reservation) => {
+  let bookingGuest = reformat.guestToName(reservation.bookingGuest);
+  let guestList = reformat.guestListToNameList(reservation.guestList);
+  let totalCost = reformat.decimal128ToMoneyString(reservation.totalCost);
+  let checkIn = reformat.toDashDate(reservation.checkIn);
+  let checkOut = reformat.toDashDate(reservation.checkOut);
+  let { _id, room_id, roomNumber = '', roomType } = reservation;
+  return {
+    bookingGuest,
+    guestList,
+    totalCost,
+    _id,
+    room_id,
+    roomNumber,
+    roomType,
+    checkIn,
+    checkOut,
+  };
+};
+
+module.exports = { formatReservation };
